Create Add_ten
1.
Create a function named add_ten that takes a dictionary with integer values named my_dictionary as a parameter. The function should add 10 to every value in my_dictionary and return my_dictionary

Loop through every key in the dictionary and add 10 to the value by using my_dictionary[key] += 10.