Create Count First Letter
Create a function named count_first_letter that takes a dictionary named names as a parameter. names should be a dictionary where the key is a last name and the value is a list of first names. For example, the dictionary might look like this:

names = {"Stark": ["Ned", "Robb", "Sansa"], "Snow" : ["Jon"], "Lannister": ["Jaime", "Cersei", "Tywin"]}
The function should return a new dictionary where each key is the first letter of a last name, and the value is the number of people whose last name begins with that letter.

So in example above, the function would return:

{"S" : 4, "L": 3}
Begin by creating an empty dictionary named something like letters. Loop through the keys of names and access the first letter of each the key using key[0].

If that letter is not a key in letters, create a new key/value pair where the key is key[0] and the value is the length of names[key].

If that letter is a key in letters, simply add the length of names[key] to value associated with key[0] in letters.

Community Forums